Make auth screen create new user by default, document complex processes

+++ app/src/main/java/com/base512/ordo/util/StringUtils.java
@@ -1,15 +1,12 @@
 package com.base512.ordo.util;
 
-import java.security.SecureRandom;
-import java.util.Random;
-
 /**
  * Created by Thomas on 2/22/2017.
  */
 
 public class StringUtils {
 
-    public static String leftZeroPad(double input, int length) {
+    public static String leftZeroPad(int input, int length) {
         return String.format("%0"+length+"d", input);
     }
 }
